Detail Screen Implemented Partially

+++ app/src/main/java/com/kelly/coinplace/domain/model/CoinSingle.kt
@@ -3,7 +3,7 @@ package com.kelly.coinplace.domain.model
 data class CoinSingle(
     val id: String,
     val athDate: String,
-    val athPrice: Int,
+    val athPrice: Double,
     val circulatingSupply: Long,
     val error: String?,
     val firstDataAt: String,
@@ -12,7 +12,7 @@ data class CoinSingle(
     val maxSupply: Long,
     val name: String,
     val price: Double,
-    val percentageChange1h: Int,
+    val percentageChange1h: Double,
     val rank: Int,
     val symbol: String,
     val totalSupply: Long
